feat: change measuring latency with int64_t

+++ src/common/monitor/stat_sum_monitor_log_write_strategy.h
@@ -10,15 +10,15 @@ namespace monitor {
 
 typedef struct ServiceInfo {
     std::string name;
-    double threshold;
+    int64_t threshold;
     std::ostream *latency_ostream;
     std::ostream *per_sec_ostream;
     std::ostream *statistic_ostream;
 
     ServiceInfo()
-        : name(""), threshold(0.0), latency_ostream(&std::cout), per_sec_ostream(&std::cout), statistic_ostream(&std::cout) {}
+        : name(""), threshold(0), latency_ostream(&std::cout), per_sec_ostream(&std::cout), statistic_ostream(&std::cout) {}
 
-    ServiceInfo(std::string n, double t, std::ostream *os1, std::ostream *os2, std::ostream *os3)
+    ServiceInfo(std::string n, int64_t t, std::ostream *os1, std::ostream *os2, std::ostream *os3)
         : name(n), threshold(t), latency_ostream(os1), per_sec_ostream(os2), statistic_ostream(os3) {}
 } ServiceInfo;
 
@@ -26,19 +26,19 @@ typedef struct ServiceInfo {
 typedef struct ProcessedLog {
     std::string content;
     std::string status;
-    double requested_at;
-    double responded_at;
-    double latency;
+    int64_t requested_at;
+    int64_t responded_at;
+    int64_t latency;
 } ProcessedLog;
 
 
 typedef struct Statistics {
     std::size_t record_cnt;
-    double epoch_size;
+    int64_t epoch_size;
     double reliability;
     double avg_latency;
-    double p90_latency;
-    double p99_latency;
+    int64_t p90_latency;
+    int64_t p99_latency;
 } Statistics;
 
 
@@ -46,10 +46,10 @@ class StatSumPerSecMonitorLogWriteStrategy : public moniq::writer::IMonitorLogWr
 private:
     std::map<std::string, ServiceInfo> services_;
     std::map<std::string, std::vector<ProcessedLog>> processed_logs_map_;
-    double epoch_size_ms_;
+    int64_t epoch_size_ms_;
 
 public:
-    StatSumPerSecMonitorLogWriteStrategy(std::vector<ServiceInfo> services, double epoch_size_ms);
+    StatSumPerSecMonitorLogWriteStrategy(std::vector<ServiceInfo> services, int64_t epoch_size_ms);
 
     ~StatSumPerSecMonitorLogWriteStrategy() = default;
